Jogadas especiais (peão para rainha)  (#114)
* pawn for queen WIP

* Change pawn to queen wip

* feat: add multiplayer (#103)

* chore(deps): bump org.sonarqube in /backend (#102)

Bumps org.sonarqube from 4.1.0.3113 to 4.2.0.3129.

---
updated-dependencies:
- dependency-name: org.sonarqube
  dependency-type: direct:production
  update-type: version-update:semver-minor
...


* chore(deps): bump element-plus from 2.3.5 to 2.3.6 in /frontend (#104)

Bumps [element-plus](https://github.com/element-plus/element-plus) from 2.3.5 to 2.3.6.
- [Release notes](https://github.com/element-plus/element-plus/releases)
- [Changelog](https://github.com/element-plus/element-plus/blob/dev/CHANGELOG.en-US.md)
- [Commits](https://github.com/element-plus/element-plus/compare/2.3.5...2.3.6)

---
updated-dependencies:
- dependency-name: element-plus
  dependency-type: direct:production
  update-type: version-update:semver-patch
...


* chore(deps-dev): bump vitest from 0.31.4 to 0.32.0 in /frontend (#106)

Bumps [vitest](https://github.com/vitest-dev/vitest/tree/HEAD/packages/vitest) from 0.31.4 to 0.32.0.
- [Release notes](https://github.com/vitest-dev/vitest/releases)
- [Commits](https://github.com/vitest-dev/vitest/commits/v0.32.0/packages/vitest)

---
updated-dependencies:
- dependency-name: vitest
  dependency-type: direct:development
  update-type: version-update:semver-minor
...


* chore(deps): bump org.sonarqube in /backend

Bumps org.sonarqube from 4.2.0.3129 to 4.2.1.3168.

---
updated-dependencies:
- dependency-name: org.sonarqube
  dependency-type: direct:production
  update-type: version-update:semver-patch
...


* Historic and the leave button (#107)

* Part of the historic

* Historic

* chore(deps): bump element-plus from 2.3.6 to 2.3.7 in /frontend

Bumps [element-plus](https://github.com/element-plus/element-plus) from 2.3.6 to 2.3.7.
- [Release notes](https://github.com/element-plus/element-plus/releases)
- [Changelog](https://github.com/element-plus/element-plus/blob/dev/CHANGELOG.en-US.md)
- [Commits](https://github.com/element-plus/element-plus/compare/2.3.6...2.3.7)

---
updated-dependencies:
- dependency-name: element-plus
  dependency-type: direct:production
  update-type: version-update:semver-patch
...


* chore(deps-dev): bump @vue/test-utils from 2.3.2 to 2.4.0 in /frontend

Bumps [@vue/test-utils](https://github.com/vuejs/test-utils) from 2.3.2 to 2.4.0.
- [Release notes](https://github.com/vuejs/test-utils/releases)
- [Commits](https://github.com/vuejs/test-utils/compare/v2.3.2...v2.4.0)

---
updated-dependencies:
- dependency-name: "@vue/test-utils"
  dependency-type: direct:development
  update-type: version-update:semver-minor
...


* Feat game score (#111)

* feat: save game score.

* test: unit tests

* chore: adjust unit test

* refactor: update name and path

* Update HomeView.vue

* chore: set-up roque

* chore: add missing }

* chore: add promotion to HORSE, TOWER, BISHOP and QUEEN

* chore: remove getPlayerService

* fix: fix mistake and add tests.

* test: add right rooke test

* refactor: simplify code

* test: add new matte test.

* feat: en passant

---------

+++ frontend/src/views/PlayerVsIaView.vue
@@ -184,12 +184,12 @@
     }
   }
 
-  function getPieceName(piece: PieceDto) {
+  function getPieceName(piece: GameMatrixPiece) {
     const pieceColor = piece.whitePiece ? "_WHITE" : "_BLACK";
     return `${piece.pieceType}${pieceColor}`;
   }
 
-  function getPieceImage(piece: PieceDto) {
+  function getPieceImage(piece: GameMatrixPiece) {
     const pieceName = getPieceName(piece);
     return pieceImages.value[pieceName];
   }
@@ -321,6 +321,7 @@
     const gameType = iaGameDto.value.gameType;
     const gameId = iaGameDto.value.id;
     const pieceId = selectedPiece.value.id;
+    const pieceType = selectedPiece.value.pieceType;
 
 
     const loading = ElLoading.service({
@@ -330,25 +331,62 @@
     });
 
     setPieceSelected(0);
+    if (move.second.first === 0 && pieceType === "PAWN") {
+      ElMessageBox.prompt("Escolha a peça para promoção, as válidas são: 'rainha', 'torre', 'cavalo', 'bispo'", "Promoção", {
+        showCancelButton: false,
+        confirmButtonText: "Confirmar",
+        inputPattern: /^(?:\s|^)(rainha|torre|cavalo|bispo)(?=\s|$)$/,
+        inputErrorMessage: "Peça inválida",
+      }).then(({ value }) => {
+        const rightValue: string = {
+          rainha: "QUEEN",
+          torre: "TOWER",
+          cavalo: "HORSE",
+          bispo: "BISHOP",
+        }[value] || "QUEEN";
 
-    GameService.movePiece(gameType, gameId, uuid, pieceId, move)
-      .then((response) => {
-        response.json().then((result) => {
-          if (result) {
-            iaGameDto.value = result;
-            localStorage.setItem("iaGameDto", JSON.stringify(result));
-            verifyGameStatus(false);
-            clearGameMatrix();
-            fillGameMatrix(result.board.whitePieces);
-            fillGameMatrix(result.board.blackPieces);
-            loading.close();
-            whiteTurn.value = false;
-          }
+        PlayerService.setPawnPromotion(playerDto.value, rightValue).then(() => {
+          GameService.movePiece(gameType, gameId, uuid, pieceId, move)
+            .then((response) => {
+              response.json().then((result) => {
+                if (result) {
+                  iaGameDto.value = result;
+                  localStorage.setItem("iaGameDto", JSON.stringify(result));
+                  verifyGameStatus(false);
+                  clearGameMatrix();
+                  fillGameMatrix(result.board.whitePieces);
+                  fillGameMatrix(result.board.blackPieces);
+                  loading.close();
+                  whiteTurn.value = false;
+                }
+              });
+            })
+            .catch(() => {
+              loading.close();
+            });
         });
-      })
-      .catch(() => {
-        loading.close();
       });
+    }
+    else {
+      GameService.movePiece(gameType, gameId, uuid, pieceId, move)
+        .then((response) => {
+          response.json().then((result) => {
+            if (result) {
+              iaGameDto.value = result;
+              localStorage.setItem("iaGameDto", JSON.stringify(result));
+              verifyGameStatus(false);
+              clearGameMatrix();
+              fillGameMatrix(result.board.whitePieces);
+              fillGameMatrix(result.board.blackPieces);
+              loading.close();
+              whiteTurn.value = false;
+            }
+          });
+        })
+        .catch(() => {
+          loading.close();
+        });
+    }
   }
 
   function getPossibleMoves(piece: GameMatrixPiece) {
